cmake: Fix PREFIX property on library targets
The `if()` command does not work inside arguments to other commands.
Simply hard-code the `PREFIX` of the static library to `lib` so that
it is called either `libjsonfortran.a` or `libjsonfortran.lib` on all
platforms.  Do not set the `PREFIX` of the shared library: they will
be called `libjsonfortran.so` or `jsonfortran.{lib,dll}` and will
not conflict with the static library.